+++ src/main/kotlin/com/dsoftware/ghtoolbar/api/Workflows.kt
@@ -47,7 +47,7 @@ object Workflows : GithubApiRequests.Entity("/repos") {
                 })
                 param(pagination)
             })
-        LOG.debug("Workflows.getWorkflowRuns() url=${url}")
+        LOG.info("Workflows.getWorkflowRuns() url=${url}")
         return get(url)
     }